chore: Divide web and mobile components in different files
