+++ src/types/types.tsx
@@ -1,0 +1,8 @@
+export interface AppBarWebProps {
+  title: string
+  version: string
+  onLogout: () => void
+}
+export interface AppBarMobileProps {
+  version: string
+}
